Rename startup modulle to plugin_registration
