+++ spyder/api/plugin_registration/decorators.py
