+++ spyder/api/plugin_registration/mixins.py
@@ -20,8 +20,8 @@ class SpyderPluginObserver:
     """
     This mixin enables a class to receive notifications when a plugin
     is available, by registering methods using the
-    :function:`spyder.api.startup.decorators.on_plugin_available` decorator,
-    which plugin to observe.
+    :function:`spyder.api.plugin_registration.decorators.on_plugin_available`
+    decorator, which plugin to observe.
 
     When any of the requested plugins is ready, the corresponding registered
     method is called.
