+++ spyder/api/plugin_registration/registry.py
