+++ spyder/api/plugins.py
@@ -40,7 +40,7 @@ from spyder.api.config.mixins import SpyderConfigurationObserver
 from spyder.api.exceptions import SpyderAPIError
 from spyder.api.plugins_enum import Plugins
 from spyder.api.translations import get_translation
-from spyder.api.startup.mixins import SpyderPluginObserver
+from spyder.api.plugin_registration.mixins import SpyderPluginObserver
 from spyder.api.widgets.main_container import PluginMainContainer
 from spyder.api.widgets.main_widget import PluginMainWidget
 from spyder.api.widgets.mixins import SpyderActionMixin
