+++ spyder/plugins/console/tests/test_plugin.py
@@ -22,7 +22,7 @@ from flaky import flaky
 # Local imports
 from spyder.api.exceptions import SpyderAPIError
 from spyder.api.plugins import SpyderPluginV2
-from spyder.api.startup.registry import PLUGIN_REGISTRY
+from spyder.api.plugin_registration.registry import PLUGIN_REGISTRY
 from spyder.config.manager import CONF
 from spyder.plugins.console.plugin import Console
 from spyder.widgets.reporterror import SpyderErrorDialog
