+++ spyder/plugins/findinfiles/plugin.py
@@ -13,7 +13,7 @@ from qtpy.QtWidgets import QApplication
 
 # Local imports
 from spyder.api.plugins import Plugins, SpyderDockablePlugin
-from spyder.api.startup.decorators import on_plugin_available
+from spyder.api.plugin_registration.decorators import on_plugin_available
 from spyder.api.translations import get_translation
 from spyder.plugins.findinfiles.widgets import FindInFilesWidget
 from spyder.plugins.mainmenu.api import ApplicationMenus
