+++ spyder/plugins/maininterpreter/tests/test_confpage.py
@@ -11,7 +11,7 @@ import time
 
 # Local imports
 from spyder.api.plugins import Plugins
-from spyder.api.startup.registry import PLUGIN_REGISTRY
+from spyder.api.plugin_registration.registry import PLUGIN_REGISTRY
 from spyder.config.manager import CONF
 from spyder.plugins.preferences.api import PreferencePages
 from spyder.plugins.maininterpreter.plugin import MainInterpreter
