+++ spyder/plugins/tours/plugin.py
@@ -12,7 +12,7 @@ Tours Plugin.
 
 # Local imports
 from spyder.api.plugins import Plugins, SpyderPluginV2
-from spyder.api.startup.decorators import on_plugin_available
+from spyder.api.plugin_registration.decorators import on_plugin_available
 from spyder.api.translations import get_translation
 from spyder.config.base import get_safe_mode, running_under_pytest
 from spyder.plugins.tours.container import ToursContainer
