+++ spyder/plugins/workingdirectory/plugin.py
@@ -16,7 +16,7 @@ from qtpy.QtCore import Signal
 
 # Local imports
 from spyder.api.plugins import SpyderPluginV2, Plugins
-from spyder.api.startup.decorators import on_plugin_available
+from spyder.api.plugin_registration.decorators import on_plugin_available
 from spyder.api.translations import get_translation
 from spyder.config.base import get_conf_path
 from spyder.plugins.workingdirectory.confpage import WorkingDirectoryConfigPage
